+++ utils/prop-types/README.md
@@ -1,0 +1,4 @@
+# @espressive/prop-types
+
+PropTypes exports all the range of validators from react prop-types that can be used to make sure the data you receive is valid. Additionally, this package provides custom prop-types that can be reusable throw all our Espressive applications, aiming to avoid duplicate code in these definitions.
+For more details, you can read react [prop-types documentation.](https://reactjs.org/docs/typechecking-with-proptypes.html)
